requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
defusedxml==0.7.1
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.13.1
httpie==3.2.1
idna==3.3
multidict==6.0.2
Pygments==2.12.0
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.10

psycopg2==2.8.3
gunicorn==19.9.0
